+++ src/components/workflow/PatientStep.tsx
@@ -1,67 +1,30 @@
-import { useState } from "react";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
-import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from "@/components/ui/dialog";
-import { Patient, Prescription } from "@/pages/Index";
+import { Patient } from "@/pages/Index";
 import { User, ArrowRight, CheckCircle } from "lucide-react";
 import PatientInformation from "@/components/PatientInformation";
 
 interface PatientStepProps {
   patients: Patient[];
-  prescriptions: Prescription[];
   selectedPatient: Patient | null;
   setSelectedPatient: (patient: Patient | null) => void;
   onAddPatient: (patient: Patient) => void;
   onUpdatePatient: (patient: Patient) => void;
   onDeletePatient: (patientId: string) => void;
   onNext: () => void;
-  onResetWorkflow: () => void;
   isCompleted: boolean;
 }
 
 const PatientStep = ({
   patients,
-  prescriptions,
   selectedPatient,
   setSelectedPatient,
   onAddPatient,
   onUpdatePatient,
   onDeletePatient,
   onNext,
-  onResetWorkflow,
   isCompleted
 }: PatientStepProps) => {
-  const [showWorkflowAlert, setShowWorkflowAlert] = useState(false);
-
-  // 진행 중인 워크플로우가 있는지 확인하는 함수
-  const hasOngoingWorkflow = (patient: Patient | null): boolean => {
-    if (!patient) return false;
-    // 해당 환자에게 prescriptions 데이터가 있는지 확인
-    const patientPrescriptions = prescriptions.filter(p => p.patientId === patient.id);
-    return patientPrescriptions.length > 0;
-  };
-
-  // TDM 선택 버튼 클릭 핸들러
-  const handleTDMSelect = () => {
-    if (hasOngoingWorkflow(selectedPatient)) {
-      setShowWorkflowAlert(true);
-    } else {
-      onNext();
-    }
-  };
-
-  // 새로 시작 버튼 클릭 핸들러
-  const handleNewStart = () => {
-    onResetWorkflow();
-    setShowWorkflowAlert(false);
-    onNext();
-  };
-
-  // 계속 진행 버튼 클릭 핸들러
-  const handleContinue = () => {
-    setShowWorkflowAlert(false);
-    onNext();
-  };
   return (
     <div className="space-y-6">
       {/* Step 1: Patient Selection */}
@@ -140,7 +103,7 @@ const PatientStep = ({
           {/* Next Button */}
           {isCompleted && (
             <div className="flex justify-end">
-              <Button onClick={handleTDMSelect} className="flex items-center gap-2 w-[300px] bg-black text-white font-bold text-lg py-3 px-6 justify-center">
+              <Button onClick={onNext} className="flex items-center gap-2 w-[300px] bg-black text-white font-bold text-lg py-3 px-6 justify-center">
                 TDM 선택
                 <ArrowRight className="h-4 w-4" />
               </Button>
@@ -148,35 +111,6 @@ const PatientStep = ({
           )}
         </CardContent>
       </Card>
-
-      {/* 워크플로우 진행 중 알림 다이얼로그 */}
-      <Dialog open={showWorkflowAlert} onOpenChange={setShowWorkflowAlert}>
-        <DialogContent className="sm:max-w-md">
-          <DialogHeader>
-            <DialogTitle className="text-lg font-bold">
-              진행 중인 TDM 분석 워크플로우가 있습니다.
-            </DialogTitle>
-          </DialogHeader>
-          <DialogDescription className="text-base">
-            {selectedPatient?.createdAt.toISOString().split('T')[0]}에 등록된 {selectedPatient?.name}환자의 워크플로우가 있습니다. 분석을 계속 진행할까요?
-          </DialogDescription>
-          <div className="flex gap-3 mt-6">
-            <Button 
-              variant="outline" 
-              onClick={handleNewStart}
-              className="flex-1"
-            >
-              새로 시작
-            </Button>
-            <Button 
-              onClick={handleContinue}
-              className="flex-1 bg-black text-white"
-            >
-              계속 진행
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
     </div>
   );
 };
